@@ -15,11 +15,11 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "spiral/logger": "^2.14.1"
+        "php": ">=8.1",
+        "spiral/logger": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5|^9.5"
+        "phpunit/phpunit": "^9.5.20"
     },
     "autoload": {
         "psr-4": {
@@ -33,7 +33,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.15.x-dev"
+            "dev-master": "2.14.x-dev"
         }
     },
     "config": {
